requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
dj-database-url==2.1.0
django==5.0.2; python_version >= '3.10'
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0; python_version >= '3.6'
gunicorn==21.2.0; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pytz==2024.1
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version >= '3.8'
whitenoise==6.6.0; python_version >= '3.8'
